+++ SRC/Injector/Public/Abstractions/IServicePath.cs
@@ -1,9 +1,9 @@
 ﻿/********************************************************************************
-* IServiceGraph.cs                                                              *
+* IServicePath.cs                                                               *
 *                                                                               *
 * Author: Denes Solti                                                           *
 ********************************************************************************/
 using Solti.Utils.DI.Interfaces;
 using System.Runtime.CompilerServices;
 
-[assembly: TypeForwardedTo(typeof(IServiceGraph))]
+[assembly: TypeForwardedTo(typeof(IServicePath))]